SEPERATE MainActivity AND LoginActivity
Modularized the authentication code, created a new login activity
that handles the login logic and modified the main activity such
that it just launches the app.
