+++ app/src/main/java/ch/epfl/sweng/radius/LoginActivity.java
@@ -1,0 +1,158 @@
+package ch.epfl.sweng.radius;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.support.annotation.NonNull;
+import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
+import android.view.View;
+import android.widget.Toast;
+
+import com.google.android.gms.auth.api.Auth;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.auth.api.signin.GoogleSignInResult;
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.SignInButton;
+import com.google.android.gms.common.api.GoogleApiClient;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.GoogleAuthProvider;
+
+import ch.epfl.sweng.radius.database.User;
+
+/**
+ *  The activity, in which the user is authenticated and logged in
+ *  using Firebase services.
+ */
+public class LoginActivity extends AppCompatActivity {
+
+    // Constants
+    private static final int RC_SIGN_IN = 1;
+    private static final String TAG = "MAIN_ACTIVITY";
+
+    // Properties
+    private SignInButton myGoogleButton;
+    private GoogleApiClient myGoogleApiClient;
+    private FirebaseAuth myAuth;
+    private FirebaseAuth.AuthStateListener myAuthListener;
+
+    public static GoogleSignInClient googleSignInClient;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        final GoogleSignInOptions googleSignInOptions = setUpGoogleSignInOptions();
+        setUpFirebaseAuthentication(googleSignInOptions);
+        setUpGoogleAPIClient(googleSignInOptions);
+        setUpGoogleSignInButton();
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        myAuth.addAuthStateListener(myAuthListener);
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        // Result returned from launching the Intent from GoogleSignInApi.getSignInIntent(...);
+        if (requestCode == RC_SIGN_IN) {
+            GoogleSignInResult result = Auth.GoogleSignInApi.getSignInResultFromIntent(data);
+            if (result.isSuccess()) {
+                //Google Sign In was successful, authentication with Firebase
+                GoogleSignInAccount account = result.getSignInAccount();
+                if (account != null) {
+                    firebaseAuthWithGoogle(account);
+                }
+            } else {
+                System.out.println("Google Sign In Failed");
+                //Google Sign In failed, update UI appropriately
+                // ...
+            }
+        }
+    }
+
+    private GoogleSignInOptions setUpGoogleSignInOptions() {
+        return new GoogleSignInOptions
+                .Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestIdToken(getString(R.string.default_web_client_id))
+                .requestEmail()
+                .build();
+    }
+
+    private void setUpFirebaseAuthentication(final GoogleSignInOptions googleSignInOptions) {
+        myAuth = FirebaseAuth.getInstance();
+        myAuthListener = new FirebaseAuth.AuthStateListener() {
+            @Override
+            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
+                if (firebaseAuth.getCurrentUser() != null) {
+                    googleSignInClient = GoogleSignIn.getClient(LoginActivity.this,
+                            googleSignInOptions);
+                    if (myAuth.getCurrentUser() != null) {
+                        User currentUser = new User(myAuth.getCurrentUser().getUid());
+                    }
+                    startActivity(new Intent(LoginActivity.this, AccountActivity.class));
+                }
+            }
+        };
+    }
+
+    private void setUpGoogleAPIClient(GoogleSignInOptions googleSignInOptions) {
+        myGoogleApiClient = new GoogleApiClient.Builder(getApplicationContext())
+                .enableAutoManage(this, new GoogleApiClient.OnConnectionFailedListener() {
+                    @Override
+                    public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+                        Toast.makeText(LoginActivity.this,
+                                "Error: Connection failed.",
+                                Toast.LENGTH_LONG).show();
+                    }
+                })
+                .addApi(Auth.GOOGLE_SIGN_IN_API, googleSignInOptions)
+                .build();
+    }
+
+    private void setUpGoogleSignInButton() {
+        SignInButton googleButton = findViewById(R.id.googleButton);
+        googleButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                signIn();
+            }
+        });
+    }
+
+    private void signIn() {
+        Intent signInIntent = Auth.GoogleSignInApi.getSignInIntent(myGoogleApiClient);
+        startActivityForResult(signInIntent, RC_SIGN_IN);
+    }
+
+    private void firebaseAuthWithGoogle(GoogleSignInAccount account) {
+        AuthCredential credential = GoogleAuthProvider.getCredential(account.getIdToken(),
+                null);
+        myAuth.signInWithCredential(credential)
+                .addOnCompleteListener(this, new OnCompleteListener<AuthResult>() {
+                    @Override
+                    public void onComplete(@NonNull Task<AuthResult> task) {
+                        Log.d(TAG, "signInWithCredential:onComplete:" + task.isSuccessful());
+
+                        //If sign in fails, display a message to the user. If sign in succeeds
+                        // the auth state listener will be notified and logic to handle the
+                        // signed in user can be handled in the listener.
+                        if (!task.isSuccessful()) {
+                            Log.w(TAG, "signInWithCredential", task.getException());
+                            Toast.makeText(LoginActivity.this, "Authentication failed.",
+                                    Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
+    }
+
+}
