+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="16dp"
     android:paddingRight="16dp"
     android:paddingTop="16dp"
-    tools:context=".MainActivity">
+    tools:context=".LoginActivity">
 
     <com.google.android.gms.common.SignInButton
         android:id="@+id/googleButton"
